+++ src/components/mentor/MentoFindSection.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useState } from "react";
 
 import EmptyMentoChatCards from "./EmptyMentoChatCards";
